refactor: change client side page version name
